:recycle: refactor: centralize test clients
