+++ src/graphql/__tests__/unit/index.test.ts
@@ -3,9 +3,9 @@ import assert from "assert";
 import { faker } from "@faker-js/faker";
 import { ResultOf } from "@graphql-typed-document-node/core";
 
-import { createMockApolloServer } from "@/graphql/__mocks__/apolloServer.js";
-import { graphql } from "@/graphql/test-utilities/unit/gql.js";
-import { GetUsersDocument } from "@/graphql/test-utilities/unit/graphql.js";
+import { createMockApolloServer } from "@/graphql/test-clients/mockApolloServer.js";
+import { graphql } from "@/graphql/test-clients/unit/gql.js";
+import { GetUsersDocument } from "@/graphql/test-clients/unit/graphql.js";
 
 describe("GraphQL", () => {
   it("server should be correctly defined", async () => {
