+++ src/graphql/test-clients/integration/index.ts
@@ -1,0 +1,1 @@
+export * from "./gql.js";
