--- src/graphql/test-utilities/integration/index.ts
@@ -1,1 +1,0 @@
-export * from "./gql.js";
