--- src/graphql/test-utilities/unit/index.ts
@@ -1,1 +1,0 @@
-export * from "./gql.js";
